// src/lib/types.gen.ts
export type Database = {
  public: {
    Tables: {
// ...
      event_tag: {
        Row: {
          event: number
          id: number
          tag: number
        }
        Insert: {
          event: number
          id?: number
          tag: number
        }
        Update: {
          event?: number
          id?: number
          tag?: number
        }
        Relationships: [
          {
            foreignKeyName: "event_tag_event_fkey"
            columns: ["event"]
            isOneToOne: false
            referencedRelation: "events"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "event_tag_tag_fkey"
            columns: ["tag"]
            isOneToOne: false
            referencedRelation: "tags"
            referencedColumns: ["id"]
          },
        ]
      }
      events: {
        Row: {
          cfpClosingDate: string | null
          cfpUrl: string | null
          city: string | null
          communityDraft: boolean | null
          communityId: number | null
          country: string | null
          date: string | null
          dateEnd: string | null
          description: string | null
          eventUrl: string | null
          id: number
          mode: Database["public"]["Enums"]["eventMode"] | null
          name: string
        }
        Insert: {
          cfpClosingDate?: string | null
          cfpUrl?: string | null
          city?: string | null
          communityDraft?: boolean | null
          communityId?: number | null
          country?: string | null
          date?: string | null
          dateEnd?: string | null
          description?: string | null
          eventUrl?: string | null
          id?: number
          mode?: Database["public"]["Enums"]["eventMode"] | null
          name: string
        }
        Update: {
          cfpClosingDate?: string | null
          cfpUrl?: string | null
          city?: string | null
          communityDraft?: boolean | null
          communityId?: number | null
          country?: string | null
          date?: string | null
          dateEnd?: string | null
          description?: string | null
          eventUrl?: string | null
          id?: number
          mode?: Database["public"]["Enums"]["eventMode"] | null
          name?: string
        }
        Relationships: [
          {
            foreignKeyName: "events_communityId_fkey"
            columns: ["communityId"]
            isOneToOne: false
            referencedRelation: "communities"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      eventMode: "In person" | "Hybrid" | "Remote"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never

export type TablesInsert<
  PublicTableNameOrOptions extends
    | keyof PublicSchema["Tables"]
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? Database[PublicTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : PublicTableNameOrOptions extends keyof PublicSchema["Tables"]
    ? PublicSchema["Tables"][PublicTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never